requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.16
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
black==23.3.0
blake3==0.3.3
blessed==1.20.0
blosc2==2.4.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.1
cycler==0.12.1
dj-database-url==2.0.0
Django==4.1.6
django-anymail==10.0
django-bootstrap5==23.3
django-clearcache==1.2.1
django-environ==0.10.0
-e git+https://github.com/jazzband/django-newsletter.git@c9d7bf2ad7cec0ffbfbfd2c0eeb0c2c8fdb66ceb#egg=django_newsletter
django-picklefield==3.1
django-q2==1.4.11
django-redis==5.2.0
docutils==0.20.1
fonttools==4.47.0
frozenlist==1.3.3
fsspec==2023.1.0
furo==2024.1.29
future==0.18.3
ghp-import==2.1.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.2.4
krock32==0.1.1
ldif3==3.1.1
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.8.2
mergedeep==1.3.4
mkdocs==1.4.3
mkdocs-material==9.1.14
mkdocs-material-extensions==1.1.1
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
ndindex==1.7
numexpr==2.8.8
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.5.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psycopg2==2.9.5
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.0
Pygments==2.15.1
pymdown-extensions==10.0.1
pyparsing==3.1.1
python-card-me==0.9.3
python-dateutil==2.8.2
pytorch-lightning==1.3.8
pytz==2022.7.1
PyYAML==5.3.1
pyyaml_env_tag==0.1
ranger21 @ git+https://github.com/lessw2020/Ranger21.git@1a96777278cdd14bc11afd865112724386d26a44
redis==4.4.2
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
sentencepiece==0.1.97
six==1.16.0
snowballstemmer==2.2.0
sorl-thumbnail==12.9.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqids==0.4.1
sqlparse==0.4.3
surlex==0.2.0
tables==3.9.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.2.0
tomli==2.0.1
torch==1.13.1
torchmetrics==0.7.0
tqdm==4.64.1
typing_extensions==4.4.0
unicodecsv==0.14.1
urllib3==1.26.14
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
whitenoise==6.5.0
yarl==1.8.2
